scripts/requirements.txt
moviepy
requests
PyYAML
